Rename package for translation service
